Update step page

+++ resources/js/views/form/index.vue
@@ -5,6 +5,7 @@ import {useToast} from 'vue-toast-notification';
 import { useRouter, useRoute } from 'vue-router';
 import { QuillEditor } from "@vueup/vue-quill";
 import draggable from "vuedraggable";
+import Swal from 'sweetalert2';
 
 const toast = useToast();
 const route = useRoute();
@@ -16,6 +17,7 @@ const footer = ref('');
 const steps = ref([]);
 const isShow = ref(false);
 const name = ref('');
+const status = ref(true);
 
 const handleSave = async () => {
     if(!form_id) {
@@ -37,7 +39,7 @@ const handleSave = async () => {
 }
 
 const handleCreate = async () => {
-    const response = await axios.post('/api/step', {name: name.value, form_id: form_id});
+    const response = await axios.post('/api/step', {name: name.value, form_id: form_id, status: status.value});
     if(response) {
         steps.value = steps.value.concat(response.data);
         isShow.value = false;
@@ -47,11 +49,23 @@ const handleCreate = async () => {
 }
 
 const handleDelete = async (id: string) => {
-    const response = await axios.delete(`/api/step/${id}`);
-    if(response) {
-        steps.value = steps.value.filter(item => item.id != id);
-        toast.success('Step deleted successfully.', {position: 'top-right'});
-    }
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+        if (result.isConfirmed) {
+            const response = await axios.delete(`/api/step/${id}`);
+            if(response) {
+                steps.value = steps.value.filter(item => item.id != id);
+            }
+            Swal.fire("Deleted!", "Step deleted successfully.", "success");
+        }
+    });
 }
 
 onMounted(async () => {
@@ -76,90 +90,102 @@ onMounted(async () => {
 </style>
 
 <template>
-    <v-row justify="end" class="mt-5">
+    <v-row justify="space-between" class="mt-5">
+        <v-btn
+            class="ml-5"
+            color="primary"
+            @click="router.push({path: '/workspace'})"
+        >
+            BACK
+        </v-btn>
         <v-btn
             class="mr-5"
-            color="blue"
+            color="primary"
             @click="handleSave()"
         >
             SAVE
         </v-btn>
     </v-row>
     <v-row>
-        <v-col cols="12">
-            <div class="font-weight-medium mb-2">
-                Header <i class="ph-asterisk ph-xs text-danger" />
-            </div>
-            <div>
-                <QuillEditor
-                    v-model:content="header"
-                    contentType="html"
-                    theme="snow"
-                />
-            </div>
-            <!-- <p>Output:</p>
-            <div v-html="header"></div> -->
+        <v-col cols="12" lg="6">
+            <Card title="Edit form" class="h-100">
+                <v-card-text>
+                    <div class="font-weight-bold mb-2 mt-3">
+                        Header <i class="ph-asterisk ph-xs text-danger" />
+                    </div>
+                    <div>
+                        <QuillEditor
+                            v-model:content="header"
+                            contentType="html"
+                            theme="snow"
+                        />
+                    </div>
+                    <div class="font-weight-bold mb-2 mt-3">
+                        Footer <i class="ph-asterisk ph-xs text-danger" />
+                    </div>
+                    <div>
+                        <QuillEditor
+                            v-model:content="footer"
+                            contentType="html"
+                            theme="snow"
+                        />
+                    </div>
+                    <!-- <p>Output:</p>
+                    <div v-html="header"></div> -->
+                </v-card-text>
+            </Card>
         </v-col>
-        <v-col cols="12">
-            <div class="font-weight-medium mb-2 mt-5">
-                Footer <i class="ph-asterisk ph-xs text-danger" />
-            </div>
-            <div>
-                <QuillEditor
-                    v-model:content="footer"
-                    contentType="html"
-                    theme="snow"
-                />
-            </div>
-        </v-col>
-        <v-col col="12" v-if="form_id">
-            <v-btn
-                @click="isShow = true"
-                color="blue"
-            >
-                Add step
-            </v-btn>
-        </v-col>
-        <v-col cols="12">
-            <v-card
-                max-width="300"
-            >
-                <v-list>
-                    <draggable v-model="steps" tag="div" item-key="id" handle=".drag-handle" animation="300">
-                        <template #item="{ element }">
-                        <v-list-item>
-                            <template v-slot:prepend>
-                            <v-icon class="drag-handle" style="cursor: grab;">mdi-drag</v-icon>
-                            </template>
-                            <v-list-item-content>
-                            <v-list-item-title>{{ element.name }}</v-list-item-title>
-                            </v-list-item-content>
-                            <template v-slot:append>
-                                <div>
-                                    <v-btn 
-                                        icon="mdi-pencil-outline" 
-                                        size="small" 
-                                        @click="router.push({path: '/step', query: {id: element.id, form_id: form_id}})">
-                                    </v-btn>
-                                    <v-btn 
-                                        icon="mdi-trash-can-outline" 
-                                        size="small" 
-                                        @click="handleDelete(element.id)">
-                                    </v-btn>
-                                </div>
-                            </template>
-                        </v-list-item>
-                        </template>
-                    </draggable>
-
-                    <!-- No Data Message inside List -->
-                    <v-list-item v-if="steps.length === 0">
-                        <v-list-item-content class="text-center">
-                        <v-list-item-title class="text-grey">No data available</v-list-item-title>
-                        </v-list-item-content>
-                    </v-list-item>
-                </v-list>
-            </v-card>
+        
+        <v-col col="12" lg="6" v-if="form_id">
+            <Card class="h-100">
+                <v-card-text>
+                    <div class="mt-4 d-flex justify-end">
+                        <v-btn
+                            @click="isShow = true"
+                            color="primary"
+                        >
+                            Add step
+                        </v-btn>
+                    </div>
+                    <div>
+                        <v-list>
+                            <draggable v-model="steps" tag="div" item-key="id" handle=".drag-handle" animation="300">
+                                <template #item="{ element }">
+                                <v-list-item>
+                                    <template v-slot:prepend>
+                                    <v-icon class="drag-handle" style="cursor: grab;">mdi-drag</v-icon>
+                                    </template>
+                                    <v-list-item-content>
+                                    <v-list-item-title>{{ element.name }}</v-list-item-title>
+                                    </v-list-item-content>
+                                    <template v-slot:append>
+                                        <div>
+                                            <v-btn 
+                                                icon="mdi-pencil-outline" 
+                                                size="small" 
+                                                @click="router.push({path: '/step', query: {id: element.id, form_id: form_id}})">
+                                            </v-btn>
+                                            <v-btn 
+                                                icon="mdi-trash-can-outline" 
+                                                size="small" 
+                                                @click="handleDelete(element.id)">
+                                            </v-btn>
+                                        </div>
+                                    </template>
+                                </v-list-item>
+                                </template>
+                            </draggable>
+                
+                            <!-- No Data Message inside List -->
+                            <v-list-item v-if="steps.length === 0">
+                                <v-list-item-content class="text-center">
+                                <v-list-item-title class="text-grey">No data available</v-list-item-title>
+                                </v-list-item-content>
+                            </v-list-item>
+                        </v-list>
+                    </div>
+                </v-card-text>
+            </Card>
         </v-col>
     </v-row>
     <v-dialog
@@ -170,20 +196,21 @@ onMounted(async () => {
             <v-card-title class="text-center">
                 <h5 class="text-h6 font-weight-bold">Add step</h5>
             </v-card-title>
-            <v-card-text class="mt-5">
-                <v-row justify="center" class="align-center">
-                    <v-col cols="12">
-                        <div class="font-weight-medium mb-2 mt-5">
-                            Step Name <i class="ph-asterisk ph-xs text-danger" />
-                        </div>
-                        <v-text-field
-                            v-model="name"
-                            isRequired
-                            placeholder="Enter step name"
-                            hide-details
-                        />
-                    </v-col>
-                </v-row>
+            <v-card-text>
+                <div class="font-weight-bold mb-2 mt-3">Name <i class="ph-asterisk ph-xs text-danger" /></div>
+                <TextField
+                    v-model="name"
+                    hide-details
+                    placeholder="Enter issue name"
+                />
+                <div class="font-weight-bold mb-2 mt-3">Status <i class="ph-asterisk ph-xs text-danger" /></div>
+                <v-switch
+                    v-model="status"
+                    color="primary"
+                    :label="status ? 'Enable' : 'Disable'"
+                    hide-details
+                    >
+                </v-switch>
             </v-card-text>
             <v-divider>
             </v-divider>
